feat: viterbi signature file

+++ src/viterbi.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
 
-void viterbi(int *obs, int *states, int *start_p, int *trans_p, int *emit_p, int T, int N, int M) {
-    
+void viterbi(float *obs, int *states, float *start_p, float *trans_p, float *emit_p, int T, int N, int M) {
+
 }
